Move .travis.yml over to travis-complex.yml default with some changes
This moves over to a prebuilt .travis.yaml with a few minor changes and
hopes to resolve #5. It also keeps stack.yaml at LTS 9.2 for the time
being and includes builds for OS X but not Windows. This does not
include:

1. Publishing haddock docs to github-pages
2. Publishing tagged versions of hackage
3. Publishing linux binaries to github-releases (on tag)
4. Publishing OS X binaries to github-releases (on tag)
5. Publishing Windows binaries to github-release (on tag)

The alternative to using the stack.yaml provided is to do 'stack init &&
stack build' but that may lead to breaking changes with certain libs
(exact-print etl. al); see #9.