Update README with requirements and installation instructions
Mac OS X seems fine (i.e. has pcre support to begin with), have tested
with Ubuntu and Fedora so package names should be correct. Have not
directly tested with NixOS / `nix`, but considering that `pcre` is the
only package that comes up with `nix-env -qaP pcre`, I'm going to assume
that's fine.

I'm not quite accustomed to cabal, so I'm not sure what to fill in below
the `stack` installation, but `stack [build|install]` seem fine when the
appropriate deps are installed.